move modules/models/utils to mmseg adapter
