+++ otx/algorithms/segmentation/adapters/mmseg/models/utils/__init__.py
@@ -1,16 +1,17 @@
+"""Utils used for mmseg model."""
+# Copyright (C) 2023 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+#
 # Copyright (c) 2020-2021 The MMSegmentation Authors
 # SPDX-License-Identifier: Apache-2.0
 #
-# Copyright (C) 2022 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-
 
 from .aggregator import IterativeAggregator, IterativeConcatAggregator
 from .angular_pw_conv import AngularPWConv
 from .asymmetric_position_attention import AsymmetricPositionAttentionModule
 from .channel_shuffle import channel_shuffle
 from .local_attention import LocalAttentionModule
+from .loss_equalizer import LossEqualizer
 from .normalize import normalize
 from .psp_layer import PSPModule
 
@@ -19,6 +20,7 @@ __all__ = [
     "IterativeConcatAggregator",
     "channel_shuffle",
     "LocalAttentionModule",
+    "LossEqualizer",
     "PSPModule",
     "AsymmetricPositionAttentionModule",
     "AngularPWConv",
