+++ otx/algorithms/segmentation/adapters/mmseg/models/utils/loss_equalizer.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 #
