--- otx/mpa/modules/models/losses/__init__.py
@@ -1,5 +1,0 @@
-# Copyright (C) 2022 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-
-# flake8: noqa
